+++ R/PDFDict.R
@@ -1,5 +1,9 @@
 
 
+create_indent <- function(depth) {
+  paste0(rep("    ", depth), collapse = "")
+}
+
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #' PDF Dictionary object creator
 #'
@@ -22,18 +26,25 @@ PDFDict <- R6::R6Class(
       invisible(self)
     },
 
+
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    # Update values in the dictionary. Each named named argument to this
+    # function is considered a name/value pair to include in the dictionary.
+    # Unnamed arguments are silently dropped
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     update = function(...) {
-      varargs <- list(...)
-      if (!all_named(varargs)) {
-        stop("PDFDict$update(): All arguments must be named")
-      }
-      self$dict <- modifyList(self$dict, varargs)
+      self$dict <- modifyList(self$dict, list(...))
       invisible(self)
     },
 
+
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    # Character representation of this dict and all sub dictionaries.
+    # @return single character string
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     as_character = function(depth = 0) {
-      indent1 <- paste0(rep("  ", depth     ), collapse = "")
-      indent2 <- paste0(rep("  ", depth + 1L), collapse = "")
+      indent1 <- create_indent(depth)
+      indent2 <- create_indent(depth + 1)
 
 
       contents <- c()
@@ -50,14 +61,26 @@ PDFDict <- R6::R6Class(
       paste0("<<\n", contents, "\n", indent1, ">>")
     },
 
+
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    # Print
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     print = function() {
       cat(self$as_character(), "\n")
     },
 
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    # Convert this dict into a full object representation for inclusion in
+    # a PDFDocument
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     as_object = function(obj_idx) {
       glue("{obj_idx} 0 obj\n{self$as_character()}\nendobj")
     },
 
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    # Since PDFDict objects can contain other PDFDict objects in a list
+    # need to do a deep copy to properly clone see private/deep_clone()
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     copy = function() {
       self$clone(deep = TRUE)
     }
@@ -69,7 +92,7 @@ PDFDict <- R6::R6Class(
     # called for every name/value pair in the object.
     # See: https://r6.r-lib.org/articles/Introduction.html
     # Need special handling for:
-    #   - 'dict' as it a list of R6 objects
+    #   - 'dict' as it a list of R6 objects or perhaps character values.
     #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     deep_clone = function(name, value) {
       if (name == 'dict') {